Add a workaround for incorrect modules.dep
Candidate releases of mangOH 0.7.0 seem to have strange dependencies
listed in modules.dep that cause the regular rather than the backported
cfg80211.ko to be loaded as a dependency of the brcmfmac.ko WiFi driver.
The root cause of why the modules.dep is wrong is unknown and there is
not enough time available to solve that issue right now, so this
workaround is being introduced until we can resolve this issue.